bindings: rust: bump system-deps
Static linking simplifies the test execution a bit (no need to set
LD_LIBRARY_PATH). system-deps gained static linking support with version
4.0, but then had a bug with it when disabling pkg-config lookup [1].
Bump the lower floor to a version with the fix.

So we can choose between 4.x or >6.1.x. In order to prevent needing to
bump the MSRV, this goes with 4.x.

API slightly changed regarding iterating the libs.

No functional changes intended. Since this is only a build-dep, no
implications should occur for users.

[1] https://github.com/gdesmott/system-deps/pull/75

Signed-off-by: Erik Schilling <erik.schilling@linaro.org>